Move prepare_static_weights_for_trtllm_fp4_moe to utils; add weight processing to w13_input_global_scale for trtllm kernel
